pyproject.toml
[tool.poetry]
name = "llm-data-analyst"
version = "0.1.0"
description = ""
authors = ["Tales Marra"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
openai = "^0.27.8"
pandas = "^2.0.3"
streamlit = "^1.25.0"
langchain = "^0.0.265"
docarray = "^0.37.0"
python-dotenv = "^1.0.0"
tabulate = "^0.9.0"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.25.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
